snapshotter unused, but compiles
