+++ mqtt/mqttd/src/broker/snapshot.rs
@@ -1,7 +1,64 @@
-use mqtt_broker::{BrokerHandle, StateSnapshotHandle};
+#![allow(dead_code)] // TODO remove when ready
+use tokio::{
+    task::JoinHandle,
+    time::{Duration, Instant},
+};
+use tracing::info;
+use tracing::warn;
+
+use mqtt_broker::{
+    BrokerHandle, FilePersistor, Message, ShutdownHandle, Snapshotter, StateSnapshotHandle,
+    SystemEvent, VersionedFileFormat,
+};
 
 pub async fn snapshot(broker_handle: BrokerHandle, snapshot_handle: StateSnapshotHandle) {
     imp::snapshot(broker_handle, snapshot_handle).await;
+}
+
+async fn start_snapshotter(
+    broker_handle: BrokerHandle,
+    persistor: FilePersistor<VersionedFileFormat>,
+    snapshot_interval: Duration,
+) -> (
+    ShutdownHandle,
+    JoinHandle<FilePersistor<VersionedFileFormat>>,
+) {
+    let snapshotter = Snapshotter::new(persistor);
+    let snapshot_handle = snapshotter.snapshot_handle();
+    let shutdown_handle = snapshotter.shutdown_handle();
+    let join_handle = tokio::spawn(snapshotter.run());
+
+    // Tick the snapshotter
+    let tick = tick_snapshot(
+        snapshot_interval,
+        broker_handle.clone(),
+        snapshot_handle.clone(),
+    );
+    tokio::spawn(tick);
+
+    // Signal the snapshotter
+    let snapshot = snapshot(broker_handle, snapshot_handle);
+    tokio::spawn(snapshot);
+
+    (shutdown_handle, join_handle)
+}
+
+async fn tick_snapshot(
+    period: Duration,
+    mut broker_handle: BrokerHandle,
+    snapshot_handle: StateSnapshotHandle,
+) {
+    info!("Persisting state every {:?}", period);
+    let start = Instant::now() + period;
+    let mut interval = tokio::time::interval_at(start, period);
+    loop {
+        interval.tick().await;
+        if let Err(e) = broker_handle.send(Message::System(SystemEvent::StateSnapshot(
+            snapshot_handle.clone(),
+        ))) {
+            warn!(message = "failed to tick the snapshotter", error=%e);
+        }
+    }
 }
 
 #[cfg(unix)]
